XPalm

Documentation for XPalm.

XPalm.AbstractPotential_EvapotranspirationModel

XPalm.AbstractRoot_GrowthModel

XPalm.AbstractSoil_WaterModel

XPalm.AbstractThermal_TimeModel

XPalm.DailyDegreeDays

XPalm.ET0_BP

XPalm.FTSW

XPalm.RootGrowth

PlantSimEngine.run!

PlantSimEngine.run!

XPalm.KS

XPalm.compute_compartment_size

XPalm.AbstractPotential_EvapotranspirationModel — Type

potential_evapotranspiration process abstract model.

All models implemented to simulate the potential_evapotranspiration process must be a subtype of this type, e.g. struct MyPotential_EvapotranspirationModel <: AbstractPotential_EvapotranspirationModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractPotential_EvapotranspirationModel)

XPalm.AbstractRoot_GrowthModel — Type

root_growth process abstract model.

All models implemented to simulate the root_growth process must be a subtype of this type, e.g. struct MyRoot_GrowthModel <: AbstractRoot_GrowthModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractRoot_GrowthModel)

XPalm.AbstractSoil_WaterModel — Type

soil_water process abstract model.

All models implemented to simulate the soil_water process must be a subtype of this type, e.g. struct MySoil_WaterModel <: AbstractSoil_WaterModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractSoil_WaterModel)

XPalm.AbstractThermal_TimeModel — Type

thermal_time process abstract model.

All models implemented to simulate the thermal_time process must be a subtype of this type, e.g. struct MyThermal_TimeModel <: AbstractThermal_TimeModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractThermal_TimeModel)

XPalm.DailyDegreeDays — Type

DailyDegreeDays(TOpt1, TOpt2, TBase, TLim)
DailyDegreeDays(TOpt1=25, TOpt2=30, TBase=15, TLim=40)

Compute thermal time from daily meteo data

Arguments

TOpt1: starting optimal temperature for thermal time calculation (degree Celsius)

TOpt2: ending optimal temperature for thermal time calculation (degree Celsius)

TBase: Tbase temperature for thermal time calculation (degree Celsius)

TLim: limit temperature for thermal time calculation (degree Celsius)

XPalm.ET0_BP — Type

ET0(LATITUDE,ALTITUDE)
ET0(LATITUDE=0.97,ALTITUDE=50)

Compute root growth depending on thermal time and water stress (ftsw)

Arguments

LATITUDE: latitude (radian)

ALTITUDE: altitude (m)

XPalm.FTSW — Type

FTSW(H_FC::Float64, H_WP_Z1::Float64,Z1::Float64,H_WP::Float64,Z2::Float64,H_0::Float64,KC::Float64,TRESH_EVAP::Float64,TRESH_FTSW_TRANSPI::Float64)

Fraction of Transpirable Soil Water model.

Arguments

H_FC: Humidity at field capacity (g[H20] g[Soil])

H_WP_Z1: Humidity at wilting point (g[H20] g[Soil]) for the first layer

Z1: Thickness of the first layer (mm)

H_WP: Humidity at wilting point (g[H20] g[Soil]) for the second layer

Z2: Thickness of the second layer (mm)

H_0: Initial soil humidity (g[H20] g[Soil])

KC: cultural coefficient (unitless)

TRESH_EVAP: fraction of water content in the evaporative layer below which evaporation is reduced (g[H20] g[Soil])

TRESH_FTSW_TRANSPI: FTSW treshold below which transpiration is reduced (g[H20] g[Soil])

XPalm.RootGrowth — Type

root_growth(TRESH_FTSW_SLOW_ROOTS, ROOTS_GROWTH_DEPTH, Z1, Z2)
root_growth(TRESH_FTSW_SLOW_ROOTS=0.2, ROOTS_GROWTH_DEPTH=0.3, Z1=200, Z2=2000)

Compute root growth depending on thermal time and water stress (ftsw)

Arguments

ROOTS_GROWTH_DEPTH: root growth in depth (mm.degreeC days-1)

TRESH_FTSW_SLOW_ROOTS: ftsw treshold below which roots growth is reduced (unitless [0:1])

Z1: Thickness of the first soil layer (mm)

Z2: Thickness of the second soil layer (mm)

PlantSimEngine.run! — Function

Compute root growth

Arguments

m: root growth model

ftsw: fraction of transpirable soil water (unitless [0:1])

TEff: daily efficient temperature for plant growth (degree C days)

Returns

root_depth: root depth (cm)

PlantSimEngine.run! — Function

Compute degree days

Arguments

m: DailyDegreeDays model

Returns

TEff: daily efficient temperature for plant growth (degree C days)

XPalm.KS — Method

KS(fillRate, tresh)

Coefficient of stress.

Arguments

fillRate: fill level of the compartment

tresh: filling treshold of the compartment below which there is a reduction in the flow

XPalm.compute_compartment_size — Method

compute_compartment_size(m, root_depth)

Compute the size of the layers of the FTSW model.

Arguments

m: FTSW model

root_depth: depth of the root system

Returns

SizeC1: size of the evapotranspirable water layer in the first soil layer (mm)

SizeVap: size of the evaporative layer within the first layer (mm)

SizeC1minusVap: size of the transpirable layer within the first layer (SizeC1-SizeVap)

SizeC2: size of the transpirable water layer in the first soil layer (mm)

SizeC: size of transpirable soil water (mm) (SizeC2 + SizeC1minusVap)