XPalm

Documentation for XPalm.

XPalm.Female

XPalm.Internode

XPalm.Leaf

XPalm.Male

XPalm.Palm

XPalm.Phytomer

XPalm.RmQ10

XPalm.add_phytomer!

XPalm.add_reproductive_organ

XPalm.determine_sex

XPalm.Female — Type

Female(state)

A female inflorescence, which has a state that can be either:

Initiation: in initiation phase (cell division)

Abortion

Flowering

Bunch: the bunch of fruits is developping

OleoSynthesis: the inflorescence is in the process of oleosynthesis

Scenescent: dead but still on the plant

Pruned: removed from the plant (e.g. harvested)

XPalm.Internode — Type

Internode(state)

An internode, which has a state of type InternodeState that can be either:

Growing: has both growth and maintenance respiration

Snag: has maintenance respiration only, and no leaf

or reproductive organs

XPalm.Leaf — Type

Leaf(state)

A leaf, which has a state of type LeafState that can be either:

Initiation: in initiation phase (cell division until begining of elongation)

Spear: spear phase, almost fully developped, but leaflets are not yet deployed

Opened: deployed and photosynthetically active

Pruned: dead and removed from the plant

Scenescent: dead but still on the plant

XPalm.Male — Type

Male(state)

A male inflorescence, which has a state that can be either:

Initiation: in initiation phase (cell division)

Abortion

Flowering

Scenescent: dead but still on the plant

Pruned: removed from the plant

XPalm.Palm — Type

Palm(mtg, phytomer_count, max_rank, node_count)
Palm()

Create a new Palm. The maximum rank is used to determine how many living phytomers (i.e. leaves) are there on the Palm.

Palm() (without arguments) creates a new Palm with a single phytomer, one leaf, and a Root system.

Arguments

mtg: a MTG object

phytomer_count: total number of phytomers emitted by the Palm since germination, i.e. physiological age

mtg_node_count: total number of nodes in the MTG (used to determine the unique ID)

XPalm.Phytomer — Type

Phytomer(state)

A phytomer

XPalm.RmQ10 — Type

RmQ10{O}(Q10) where O <: Organ

Maintenance respiration of an organ.

Examples

RmQ10{Leaf}(1.5)

XPalm.add_phytomer! — Method

add_phytomer!(palm, initiation_date)

Add a new phytomer to the palm

Arguments

palm: a Palm

initiation_date::Dates.Date: date of initiation of the phytomer

XPalm.add_reproductive_organ — Method

ex: add_reproductive_organ(node[:organ], node)

XPalm.determine_sex — Method

Determine the sex of the reproductive organ based on the trophic state of the palm tree on x last days