XPalm

Documentation for XPalm.

XPalm.AbstractCarbon_AllocationModel

XPalm.AbstractCarbon_AssimilationModel

XPalm.AbstractCarbon_DemandModel

XPalm.AbstractCarbon_OfferModel

XPalm.AbstractFTSWModel

XPalm.AbstractInitiation_AgeModel

XPalm.AbstractLai_DynamicModel

XPalm.AbstractLeaf_AreaModel

XPalm.AbstractLeaf_Final_Potential_AreaModel

XPalm.AbstractLeaf_Potential_AreaModel

XPalm.AbstractLeaf_PruningModel

XPalm.AbstractLeaf_RankModel

XPalm.AbstractLight_InterceptionModel

XPalm.AbstractMaintenance_RespirationModel

XPalm.AbstractPhyllochronModel

XPalm.AbstractPhytomer_EmissionModel

XPalm.AbstractPlant_AgeModel

XPalm.AbstractPotential_EvapotranspirationModel

XPalm.AbstractRoot_GrowthModel

XPalm.AbstractSoil_WaterModel

XPalm.AbstractStateModel

XPalm.AbstractThermal_TimeModel

XPalm.Beer

XPalm.CarbonOfferPhotosynthesis

XPalm.ConstantRUEModel

XPalm.DailyDegreeDays

XPalm.DailyPlantAgeModel

XPalm.DegreeDaysFTSW

XPalm.ET0_BP

XPalm.FTSW

XPalm.FTSW_BP

XPalm.Female

XPalm.Internode

XPalm.Leaf

XPalm.LeafCarbonDemandModel

XPalm.Male

XPalm.Palm

XPalm.Phytomer

XPalm.PotentialAreaModel

XPalm.RmQ10

XPalm.RootGrowthFTSW

PlantSimEngine.run!

PlantSimEngine.run!

PlantSimEngine.run!

PlantSimEngine.run!

PlantSimEngine.run!

XPalm.KS

XPalm.KS_bp

XPalm.compute_compartment_size

XPalm.compute_compartment_size

XPalm.AbstractCarbon_AllocationModel — Type

carbon_allocation process abstract model.

All models implemented to simulate the carbon_allocation process must be a subtype of this type, e.g. struct MyCarbon_AllocationModel <: AbstractCarbon_AllocationModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractCarbon_AllocationModel)

XPalm.AbstractCarbon_AssimilationModel — Type

carbon_assimilation process abstract model.

All models implemented to simulate the carbon_assimilation process must be a subtype of this type, e.g. struct MyCarbon_AssimilationModel <: AbstractCarbon_AssimilationModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractCarbon_AssimilationModel)

XPalm.AbstractCarbon_DemandModel — Type

carbon_demand process abstract model.

All models implemented to simulate the carbon_demand process must be a subtype of this type, e.g. struct MyCarbon_DemandModel <: AbstractCarbon_DemandModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractCarbon_DemandModel)

XPalm.AbstractCarbon_OfferModel — Type

carbon_offer process abstract model.

All models implemented to simulate the carbon_offer process must be a subtype of this type, e.g. struct MyCarbon_OfferModel <: AbstractCarbon_OfferModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractCarbon_OfferModel)

XPalm.AbstractFTSWModel — Type

AbstractFTSWModel <: AbstractSoil_WaterModel

Defines a structure for soil water model that computes ftsw as an output.

XPalm.AbstractInitiation_AgeModel — Type

initiation_age process abstract model.

All models implemented to simulate the initiation_age process must be a subtype of this type, e.g. struct MyInitiation_AgeModel <: AbstractInitiation_AgeModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractInitiation_AgeModel)

XPalm.AbstractLai_DynamicModel — Type

lai_dynamic process abstract model.

All models implemented to simulate the lai_dynamic process must be a subtype of this type, e.g. struct MyLai_DynamicModel <: AbstractLai_DynamicModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractLai_DynamicModel)

XPalm.AbstractLeaf_AreaModel — Type

leaf_area process abstract model.

All models implemented to simulate the leaf_area process must be a subtype of this type, e.g. struct MyLeaf_AreaModel <: AbstractLeaf_AreaModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractLeaf_AreaModel)

XPalm.AbstractLeaf_Final_Potential_AreaModel — Type

leaf_final_potential_area process abstract model.

All models implemented to simulate the leaf_final_potential_area process must be a subtype of this type, e.g. struct MyLeaf_Final_Potential_AreaModel <: AbstractLeaf_Final_Potential_AreaModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractLeaf_Final_Potential_AreaModel)

XPalm.AbstractLeaf_Potential_AreaModel — Type

leaf_potential_area process abstract model.

All models implemented to simulate the leaf_potential_area process must be a subtype of this type, e.g. struct MyLeaf_Potential_AreaModel <: AbstractLeaf_Potential_AreaModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractLeaf_Potential_AreaModel)

XPalm.AbstractLeaf_PruningModel — Type

leaf_pruning process abstract model.

All models implemented to simulate the leaf_pruning process must be a subtype of this type, e.g. struct MyLeaf_PruningModel <: AbstractLeaf_PruningModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractLeaf_PruningModel)

XPalm.AbstractLeaf_RankModel — Type

leaf_rank process abstract model.

All models implemented to simulate the leaf_rank process must be a subtype of this type, e.g. struct MyLeaf_RankModel <: AbstractLeaf_RankModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractLeaf_RankModel)

XPalm.AbstractLight_InterceptionModel — Type

light_interception process abstract model.

All models implemented to simulate the light_interception process must be a subtype of this type, e.g. struct MyLight_InterceptionModel <: AbstractLight_InterceptionModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractLight_InterceptionModel)

XPalm.AbstractMaintenance_RespirationModel — Type

maintenance_respiration process abstract model.

All models implemented to simulate the maintenance_respiration process must be a subtype of this type, e.g. struct MyMaintenance_RespirationModel <: AbstractMaintenance_RespirationModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractMaintenance_RespirationModel)

Generic maintenance repspiration model.

The models used are defined by the types of the maintenance_respiration fields of a ModelList.

Examples

meteo = Atmosphere(T = 20.0, Wind = 1.0, P = 101.3, Rh = 0.65)

# Using Fvcb model:
leaf =
    ModelList(
        maintenance_respiration = RmQ10(),
        status = (Tₗ = 25.0, PPFD = 1000.0, Cₛ = 400.0, Dₗ = meteo.VPD)
    )

maintenance_respiration(leaf, meteo)

XPalm.AbstractPhyllochronModel — Type

phyllochron process abstract model.

All models implemented to simulate the phyllochron process must be a subtype of this type, e.g. struct MyPhyllochronModel <: AbstractPhyllochronModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractPhyllochronModel)

XPalm.AbstractPhytomer_EmissionModel — Type

phytomer_emission process abstract model.

All models implemented to simulate the phytomer_emission process must be a subtype of this type, e.g. struct MyPhytomer_EmissionModel <: AbstractPhytomer_EmissionModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractPhytomer_EmissionModel)

XPalm.AbstractPlant_AgeModel — Type

plant_age process abstract model.

All models implemented to simulate the plant_age process must be a subtype of this type, e.g. struct MyPlant_AgeModel <: AbstractPlant_AgeModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractPlant_AgeModel)

XPalm.AbstractPotential_EvapotranspirationModel — Type

potential_evapotranspiration process abstract model.

All models implemented to simulate the potential_evapotranspiration process must be a subtype of this type, e.g. struct MyPotential_EvapotranspirationModel <: AbstractPotential_EvapotranspirationModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractPotential_EvapotranspirationModel)

XPalm.AbstractRoot_GrowthModel — Type

root_growth process abstract model.

All models implemented to simulate the root_growth process must be a subtype of this type, e.g. struct MyRoot_GrowthModel <: AbstractRoot_GrowthModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractRoot_GrowthModel)

XPalm.AbstractSoil_WaterModel — Type

soil_water process abstract model.

All models implemented to simulate the soil_water process must be a subtype of this type, e.g. struct MySoil_WaterModel <: AbstractSoil_WaterModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractSoil_WaterModel)

XPalm.AbstractStateModel — Type

state process abstract model.

All models implemented to simulate the state process must be a subtype of this type, e.g. struct MyStateModel <: AbstractStateModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractStateModel)

XPalm.AbstractThermal_TimeModel — Type

thermal_time process abstract model.

All models implemented to simulate the thermal_time process must be a subtype of this type, e.g. struct MyThermal_TimeModel <: AbstractThermal_TimeModel end.

You can list all models implementing this process using subtypes:

Examples

subtypes(AbstractThermal_TimeModel)

XPalm.Beer — Type

Beer(k)

Beer-Lambert law for light interception.

Required inputs: LAI in m² m⁻². Required meteorology data: Ri_PAR_f, the incident flux of atmospheric radiation in the PAR, in W m[soil]⁻² (== J m[soil]⁻² s⁻¹).

Output: aPPFD, the absorbed Photosynthetic Photon Flux Density in μmol[PAR] m[leaf]⁻² s⁻¹.

XPalm.CarbonOfferPhotosynthesis — Type

CarbonOfferPhotosynthesis()

A model that computes carbon offer as the carbon_assimilation.

XPalm.ConstantRUEModel — Type

ConstantRUEModel(rue)

Computes the carbon_assimilation using a constant radiation use efficiency (rue).

XPalm.DailyDegreeDays — Type

DailyDegreeDays(TOpt1, TOpt2, TBase, TLim)
DailyDegreeDays(TOpt1=25, TOpt2=30, TBase=15, TLim=40)

Compute thermal time from daily meteo data

Arguments

TOpt1: starting optimal temperature for thermal time calculation (degree Celsius)

TOpt2: ending optimal temperature for thermal time calculation (degree Celsius)

TBase: Tbase temperature for thermal time calculation (degree Celsius)

TLim: limit temperature for thermal time calculation (degree Celsius)

XPalm.DailyPlantAgeModel — Type

DailyPlantAgeModel(initiation_age)

Plant age model, simply tracks the age of the plant in days.

Arguments

initiation_age: age of the plant at the start of the simulation (days)

Returns

age: age of the plant (days)

XPalm.DegreeDaysFTSW — Type

DegreeDaysFTSW(TOpt1, TOpt2, TBase, TLim, threshold_ftsw_stress)
DegreeDaysFTSW(TOpt1=25, TOpt2=30, TBase=15, TLim=40, threshold_ftsw_stress=0.3)

Compute thermal time from daily meteo data, corrected by FTSW

Arguments

TOpt1: starting optimal temperature for thermal time calculation (degree Celsius)

TOpt2: ending optimal temperature for thermal time calculation (degree Celsius)

TBase: Tbase temperature for thermal time calculation (degree Celsius)

TLim: limit temperature for thermal time calculation (degree Celsius)

threshold_ftsw_stress: threshold value under which we apply an FTSW stress

XPalm.ET0_BP — Type

ET0_BP(LATITUDE,ALTITUDE)
ET0_BP(LATITUDE=0.97,ALTITUDE=50)

Compute root growth depending on thermal time and water stress (ftsw)

Arguments

LATITUDE: latitude (radian)

ALTITUDE: altitude (m)

XPalm.FTSW — Type

FTSW{O}(
    ini_root_depth::T
    H_FC::T
    H_WP_Z1::T
    Z1::T
    H_WP_Z2::T
    Z2::T
    H_0::T
    KC::T
    TRESH_EVAP::T
    TRESH_FTSW_TRANSPI::T
    ini_qty_H2O_Vap::T
    ini_qty_H2O_C1::T
    ini_qty_H2O_C1minusVap::T
    ini_qty_H2O_C2::T
    ini_qty_H2O_C::T
)

Fraction of Transpirable Soil Water model.

Note that there is also a method for FTSW that takes an organ type as type, e.g. FTSW{Leaf}(ini_root_depth = 200.0).

Arguments

ini_root_depth: root depth at initialization (mm)

H_FC: Humidity at field capacity (g[H20] g[Soil])

H_WP_Z1: Humidity at wilting point (g[H20] g[Soil]) for the first layer

Z1: Thickness of the first layer (mm)

H_WP_Z2: Humidity at wilting point (g[H20] g[Soil]) for the second layer

Z2: Thickness of the second layer (mm)

H_0: Initial soil humidity (g[H20] g[Soil])

KC: cultural coefficient (unitless)

TRESH_EVAP: fraction of water content in the evaporative layer below which evaporation is reduced (g[H20] g[Soil])

TRESH_FTSW_TRANSPI: FTSW treshold below which transpiration is reduced (g[H20] g[Soil])

XPalm.FTSW_BP — Type

FTSW_BP(H_FC::Float64, H_WP_Z1::Float64,Z1::Float64,H_WP_Z2::Float64,Z2::Float64,H_0::Float64,KC::Float64,TRESH_EVAP::Float64,TRESH_FTSW_TRANSPI::Float64)

Fraction of Transpirable Soil Water model.

Arguments

ini_root_depth: root depth at initialization (mm)

H_FC: Humidity at field capacity (g[H20] g[Soil])

H_WP_Z1: Humidity at wilting point (g[H20] g[Soil]) for the first layer

Z1: Thickness of the first layer (mm)

H_WP_Z2: Humidity at wilting point (g[H20] g[Soil]) for the second layer

Z2: Thickness of the second layer (mm)

H_0: Initial soil humidity (g[H20] g[Soil])

KC: cultural coefficient (unitless)

TRESH_EVAP: fraction of water content in the evaporative layer below which evaporation is reduced (g[H20] g[Soil])

TRESH_FTSW_TRANSPI: FTSW treshold below which transpiration is reduced (g[H20] g[Soil])

XPalm.Female — Type

Female(state)

A female inflorescence, which has a state that can be either:

Initiation: in initiation phase (cell division)

Abortion

Flowering

Bunch: the bunch of fruits is developping

OleoSynthesis: the inflorescence is in the process of oleosynthesis

Scenescent: dead but still on the plant

Pruned: removed from the plant (e.g. harvested)

XPalm.Internode — Type

Internode(state)

An internode, which has a state of type InternodeState that can be either:

Growing: has both growth and maintenance respiration

Snag: has maintenance respiration only, and no leaf

or reproductive organs

XPalm.Leaf — Type

Leaf(state)

A leaf, which has a state of type LeafState that can be either:

Initiation: in initiation phase (cell division until begining of elongation)

Spear: spear phase, almost fully developped, but leaflets are not yet deployed

Opened: deployed and photosynthetically active

Pruned: dead and removed from the plant

Scenescent: dead but still on the plant

XPalm.LeafCarbonDemandModel — Type

LeafCarbonDemandModel(lma_min, respiration_cost, leaflets_biomass_contribution)

Carbon demand of the leaf based on the potential leaf area increment of the day.

Arguments

lma_min: minimum leaf mass area (g m⁻²)

respiration_cost: growth respiration cost (g g⁻¹)

leaflets_biomass_contribution: contribution of the leaflet biomass to the total leaf biomass (including rachis)

XPalm.Male — Type

Male(state)

A male inflorescence, which has a state that can be either:

Initiation: in initiation phase (cell division)

Abortion

Flowering

Scenescent: dead but still on the plant

Pruned: removed from the plant

XPalm.Palm — Type

Palm(;
    nsteps=1,
    initiation_age=0,
    parameters=default_parameters(),
    model_list=main_models_definition(parameters, nsteps)
)

Create a new scene with one Palm plant.

Arguments

nsteps: number of time steps to run the simulation for (default: 1, should match the number of rows in the meteo data)

initiation_age: date of the first phytomer initiation (default: 0)

parameters: a dictionary of parameters (default: default_parameters())

model_list: a dictionary of models (default: main_models_definition(parameters, nsteps))

XPalm.Phytomer — Type

Phytomer(state)

A phytomer

XPalm.PotentialAreaModel — Type

PotentialAreaModel(inflexion_index, slope)

Computes the instantaneous potential area at a given cumulative thermal time using a logistic function. In other words, it defines the development of the leaf area at the potential, i.e. without any stress. It starts around 0.0 and goes to a maximum of final_potential_area.

Arguments

inflexion_index: a parameter that defines the relationship between the final potential

leaf area and the inflexion point of the logistic function. The higher the final area, the longer the time to reach the inflexion point.

slope: the slope of the relationship at the inflexion point.

Inputs

The model needs two input variables:

final_potential_area: the final potential area when the leaf is fully developed

TT_since_init: the cumulated thermal time since leaf initiation

XPalm.RmQ10 — Type

RmQ10{O}(Q10) where O <: Organ

Maintenance respiration.

Arguments

Q10: Q10 factor (values should usually range between: 1.5 - 2.5, with 2.1 being the most common value)

Rm_base: Base maintenance respiration (gC gDM⁻¹ d⁻¹). Should be around 0.06.

T_ref: Reference temperature at which Q10 was measured (default: 25.0°C)

Examples

RmQ10{Leaf}(1.5, 0.06, 25.0)

XPalm.RootGrowthFTSW — Type

root_growth(TRESH_FTSW_SLOW_ROOTS, ROOTS_GROWTH_DEPTH, Z1, Z2)
root_growth(TRESH_FTSW_SLOW_ROOTS=0.2, ROOTS_GROWTH_DEPTH=0.3, Z1=200, Z2=2000)

Compute root growth depending on thermal time and water stress (ftsw)

Arguments

ini_root_depth: initial root depth (mm)

ROOTS_GROWTH_DEPTH: root growth in depth (mm.degreeC days-1)

TRESH_FTSW_SLOW_ROOTS: ftsw treshold below which roots growth is reduced (unitless [0:1])

The model as a dependency on an AbstractFTSWModel that must return a value for ftsw, and soil_depth.

PlantSimEngine.run! — Function

Compute root growth

Arguments

m: root growth model

ftsw: fraction of transpirable soil water (unitless [0:1])

TEff: daily efficient temperature for plant growth (degree C days)

Returns

root_depth: root depth (cm)

PlantSimEngine.run! — Function

Compute degree days

Arguments

m: DailyDegreeDays model

Returns

TEff: daily efficient temperature for plant growth (degree C days)

PlantSimEngine.run! — Function

Compute degree days corrected by FTSW

Arguments

m: DegreeDaysFTSW model

Returns

TEff: daily efficient temperature for plant growth (degree C days)

PlantSimEngine.run! — Function

run!(object, meteo, constants = Constants())

Computes the light interception of an object using the Beer-Lambert law.

Arguments

::Beer: a Beer model, from the model list (i.e. m.light_interception)

models: A ModelList struct holding the parameters for the model with

initialisations for lai (m² m⁻²): the leaf area index.

status: the status of the model, usually the model list status (i.e. m.status)

meteo: meteorology structure, see Atmosphere

constants = PlantMeteo.Constants(): physical constants. See PlantMeteo.Constants for more details

Examples

using PlantSimEngine, PlantBiophysics, PlantMeteo
m = ModelList(light_interception=Beer(0.5), status=(LAI=2.0,))

meteo = Atmosphere(T=20.0, Wind=1.0, P=101.3, Rh=0.65, Ri_PAR_f=300.0)

run!(m, meteo)

m[:aPPFD]

PlantSimEngine.run! — Method

add_phytomer!(palm, initiation_age)

Add a new phytomer to the palm

Arguments

palm: a Palm

initiation_age::Dates.Date: date of initiation of the phytomer

XPalm.KS — Method

KS(fillRate, tresh)

Coefficient of stress.

Arguments

fillRate: fill level of the compartment

tresh: filling treshold of the compartment below which there is a reduction in the flow

XPalm.KS_bp — Method

KS_bp(fillRate, tresh)

Coefficient of stress.

Arguments

fillRate: fill level of the compartment

tresh: filling treshold of the compartment below which there is a reduction in the flow

XPalm.compute_compartment_size — Method

compute_compartment_size(m, root_depth)

Compute the size of the layers of the FTSW model.

Arguments

m: FTSW model

root_depth: depth of the root system

Returns

SizeC1: size of the evapotranspirable water layer in the first soil layer (mm)

SizeVap: size of the evaporative layer within the first layer (mm)

SizeC1minusVap: size of the transpirable layer within the first layer (SizeC1-SizeVap)

SizeC2: size of the transpirable water layer in the first soil layer (mm)

SizeC: size of transpirable soil water (mm) (SizeC2 + SizeC1minusVap)

XPalm.compute_compartment_size — Method

compute_compartment_size(m, root_depth)

Compute the size of the layers of the FTSW model.

Arguments

m: FTSW model

root_depth: depth of the root system

Returns

SizeC1: size of the evapotranspirable water layer in the first soil layer (mm)

SizeVap: size of the evaporative layer within the first layer (mm)

SizeC1minusVap: size of the transpirable layer within the first layer (SizeC1-SizeVap)

SizeC2: size of the transpirable water layer in the first soil layer (mm)

SizeC: size of transpirable soil water (mm) (SizeC2 + SizeC1minusVap)